moved page components into seperate folders
