+++ tech-view/src/App.js
@@ -1,10 +1,11 @@
 import HomePageWithoutLogin from './components/homePageWithoutlogin/HomePageWithoutLogin'
-import MainPage from './components/MainPage'
+import MainPage from './components/mainPageWithLogin/MainPage'
+// import MainPage from './components/MainPage'
 
 function App () {
   return (
-    // <MainPage />
     <HomePageWithoutLogin />
+    // <MainPage />
   )
 }
