+++ tech-view/src/components/homePageWithoutlogin/TopPortionOfHomePage.js
@@ -1,6 +1,6 @@
 export default function TopPortionOfHomePage () {
   return (
-    <div style={{ marginTop: '89px', display: 'flex', justifyContent: 'space-between' }}>
+    <div style={{ marginTop: '89px', padding: '30px', display: 'flex', justifyContent: 'space-between' }}>
       <h2>LOREM IPSUM DOLOR</h2>
       <img alt='people collaborating' />
     </div>
