+++ tech-view/src/components/mainPageWithLogin/LikeAndCommentButtonContainer.js
