+++ tech-view/src/components/mainPageWithLogin/MainPage.js
@@ -1,7 +1,7 @@
-import Navbar from './Navbar'
+import Navbar from '../Navbar'
 import PostCardsContainer from './PostCardsContainer'
-import SideBar from './SideBar'
-import Footer from './Footer'
+import SideBar from '../SideBar'
+import Footer from '../Footer'
 
 export default function MainPage () {
   return (
