+++ tech-view/src/components/mainPageWithLogin/PostCard.js
