+++ tech-view/src/components/mainPageWithLogin/PostInfo.js
